wire up ProjectMembership and Person to be part of generators
