+++ pyvo/model/person.py
@@ -1,4 +1,4 @@
-from base import PivotalResource, Instantiated, fields
+from base import PivotalResource, Instantiated, fields, OneOf
 from metadata import TimeZone
 from project import MembershipSummary
 
@@ -24,3 +24,20 @@ class Person(Instantiated, PivotalResource):
     initials = fields.StringField()
     username = fields.StringField()
     kind = fields.StringField()
+
+class ProjectMembership(Instantiated, PivotalResource):
+
+    created_at = fields.DateTimeField()
+    updated_at = fields.DateTimeField()
+    person_id = fields.IntField()
+    project_id = fields.IntField()
+    role = fields.StringField(
+        validators=OneOf('owner',
+            'member', 'viewer', 'inactive'))
+    project_color = fields.StringField()
+    favorite = fields.BoolField()
+    last_viewed_at = fields.DateTimeField()
+    wants_comment_notification_emails = fields.BoolField()
+    will_receive_mention_notifications_or_emails = fields.BoolField()
+    kind = fields.StringField()
+    person = fields.EmbeddedField(Person)
